// src/types/product.ts
export interface ProductExtra {
  id: string;
  name: string;
  price: number;
}

export interface Product {
  id: string;
  name: string;
  description: string;
  price: number;
  originalPrice?: number;
  image: string;
  category: string;
  isHighlight?: boolean;
  is_available?: boolean;
  extras?: ProductExtra[];
}